+++ client/pages/performances/_id.vue
@@ -1,0 +1,12 @@
+<template>
+  <h1>{{ id }}</h1>
+</template>
+
+<script>
+export default ({
+  asyncData ({ params }) {
+    const id = params.id
+    return { id }
+  }
+})
+</script>
